task-6/event-producer/requirements.txt
flask==3.0.3
waitress==3.0.2
kafka-python==2.1.0
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-instrumentation-flask==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-instrumentation-kafka-python==0.58b0
